feat: Divide api modules by feature group
